move kernel spec to its own file; add xdg-portable
