+++ extensions/positron-r/package.json
@@ -674,7 +674,8 @@
     "split2": "^4.2.0",
     "vscode-languageclient": "^9.0.1",
     "web-tree-sitter": "^0.20.8",
-    "which": "^3.0.0"
+    "which": "^3.0.0",
+    "xdg-portable": "^10.6.0"
   },
   "peerDependencies": {
     "@vscode/windows-registry": "^1.0.0"
