+++ extensions/positron-r/package.nls.json
@@ -61,8 +61,8 @@
 	"r.configuration.diagnostics.enable.description": "Enable R diagnostics globally",
 	"r.configuration.taskHyperlinks.description": "Turn on experimental support for hyperlinks in package development tasks",
 	"r.configuration.defaultRepositories.description": "The default repositories to use for R package installation, if no repository is otherwise specified in R startup scripts (restart Positron to apply).\n\nThe default repositories will be set as the `repos` option in R.",
-	"r.configuration.defaultRepositories.auto.description": "Automatically choose a default repository, or use a .conf file if it exists.",
+	"r.configuration.defaultRepositories.auto.description": "Automatically choose a default repository, or use a repos.conf file if it exists.",
 	"r.configuration.defaultRepositories.rstudio.description": "Use the RStudio CRAN mirror (cran.rstudio.com)",
-	"r.configuration.defaultRepositories.posit-ppm.description": "Use the RStudio public package manager (packagemanager.rstudio.com)",
+	"r.configuration.defaultRepositories.posit-ppm.description": "Use the Posit public package manager (packagemanager.posit.co)",
 	"r.configuration.defaultRepositories.none.description": "Do not set a default repository or change the value of the 'repos' option"
 }
